Attempt to speed up Delaunay triangulation
ATM: Deaunay Triangulation sorts points around orgin every time it merges
two smaller triangulations. It can be easily made faster if neighbours were
kept sorted around origin. I tried to do it, but they have to be kept in
a set, so the set of neighbours of a vertex v should be able to sort around
v, so it needs to know v - thus lambdas with capturing. It does not complie.